Create Home Screen.
