+++ lib/main.dart
@@ -2,9 +2,12 @@ import 'package:firebase_core/firebase_core.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:go_router/go_router.dart';
-import 'package:social_media/screens/home.dart';
-import 'package:social_media/screens/sign_in.dart';
-import 'package:social_media/screens/sign_up.dart';
+import 'package:social_media/presentation/home.dart';
+import 'package:social_media/presentation/auth/sign_in.dart';
+import 'package:social_media/presentation/auth/sign_up.dart';
+import 'package:social_media/presentation/splash.dart';
+import 'package:social_media/utils/responsive.dart';
+import 'package:social_media/utils/colors.dart';
 
 void main() async {
   WidgetsFlutterBinding.ensureInitialized();
@@ -23,7 +26,13 @@ class MyApp extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
+    AppResponsive.init(context);
     return MaterialApp.router(
+      theme: ThemeData(
+        splashFactory: NoSplash.splashFactory,
+        splashColor: AppColors.grey,
+        highlightColor: AppColors.grey,
+      ),
       routerConfig: _router,
       debugShowCheckedModeBanner: false,
     );
@@ -31,8 +40,24 @@ class MyApp extends StatelessWidget {
 }
 
 final _router = GoRouter(
-  initialLocation: "/sign_in",
+  initialLocation: "/",
   routes: [
+    GoRoute(
+      path: "/",
+      pageBuilder: (context, state) {
+        return CustomTransitionPage(
+          key: state.pageKey,
+          transitionDuration: const Duration(milliseconds: 300),
+          child: SplashScreen(),
+          transitionsBuilder: (context, animation, secondaryAnimation, child) {
+            return FadeTransition(
+              opacity: CurveTween(curve: Curves.easeIn).animate(animation),
+              child: child,
+            );
+          },
+        );
+      },
+    ),
     GoRoute(
       path: "/sign_in",
       pageBuilder: (context, state) {
