+++ lib/utils/colors.dart
@@ -1,8 +1,8 @@
 import 'package:flutter/material.dart';
 
-class AppColors{
+class AppColors {
   static Color black = Colors.black;
   static Color white = Colors.white;
   static Color grey = Colors.grey;
-  static Color brown = Colors.brown;
+  static Color brown = Colors.brown[700]!;
 }
