+++ pubspec.yaml
@@ -26,6 +26,9 @@ dependencies:
   flutter_spinkit: ^5.2.1
   logger: ^2.5.0
   shared_preferences: ^2.3.4
+  iconsax_plus: ^1.0.0
+  iconsax: ^0.0.8
+  ionicons: ^0.2.2
